requirements.txt
playwright==1.50.0
playwright-stealth==2.0.2
python-dotenv==1.0.0
pytz==2025.1
Pillow==11.0.0
